backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
numpy>=1.24.0
Pillow>=9.0.0
joblib>=1.2.0
python-multipart>=0.0.6
tensorflow>=2.13.0
scikit-learn>=1.3.0
opencv-python-headless>=4.8.0
